Retry failed ML Job setup in DE FTR tests
The flakiness here ends up being caused by sporadic unavailability of
shards during module setup. The underlying cause of that unavailability
is likely a race condition between ML, ES, and/or FTR, but luckily we
don't need to worry about that because simply retrying the API call
causes it to eventually succeed.

In those cases, some of the jobs will report a 4xx status, but that's
expected.

This is the result of a lot of prodding and CPU cycles on CI; see #182183 for
the full details.