[Bug][Investigations] - Fix timeline column reset (#188240)
## Summary

**Bug:**
Within the new timeline, we have default columns that show whenever an
empty array for columns are provided.
This was causing columns to re-appear in the `Selected fields` section
even though the columns were not actually selected anymore leading to
mismatched state between the browser ui and the table preventing the
default columns from being removed again without page refresh. _The old
timeline table is not affected_.

https://github.com/user-attachments/assets/7d16b8c3-be1a-4bc7-87b0-62fcac563f15

**Fix:**
The logic used to default an empty array of columns to the default
timeline columns has been updated to allow the empty array to be passed.
With the old timeline experience this just means all columns _can_ be
removed and you'd see an empty table, but with the new unified table
(which will be the default in 8.15 after

https://github.com/user-attachments/assets/e5fbaf48-77e3-4363-8232-35d256729b7e
(cherry picked from commit 385884d98b21e5613eadb613ef984aa87da282ae)

# Conflicts:
#	x-pack/plugins/security_solution/public/timelines/components/timeline/tabs/shared/__snapshots__/use_timeline_columns.test.ts.snap