Tag serverless UI tests for esGate (#188239)
## Summary

This PR tags the majority of serverless FTR UI tests with `esGate` so
they're included in the elasticsearch-serverless commit verification
process.

### Details

#### Why we do this

We had cases where elasticsearch-serverless changes broke Kibana
functionality and we only discovered this very late in the cycle, in
some cases even after release to production. An investigation showed
that these cases would have been caught by the Kibana serverless UI
tests. So this is added in the interest of closing gaps in ES-KBN
compatibility checks.

#### Stability and run time

Run time:
One requirement for adding more tests to the elasticsearch-serverless
verification is that the overall run time is not increased. When
comparing the ES verification run time details from the ES process with
the ES verification from the Kibana process (daily run where we already
include UI tests), we can see that the overall run time is dominated by
Jest integration tests. So enabling the UI tests will not extend the
overall run time for the ES process.

Test stability:
Initially, we didn't include UI tests in the elasticsearch-serverless
verification because of test flakiness concerns. This has greatly
improved since the early days of serverless tests and we only rarely see
UI test failures on `main` and if we do, they're addressed quickly (the
chart shows number of serverless UI test failures on `main` over the
past 30 days). We do not run the ES verification against latest `main`
but only against a successfully checked commit level, so the chance that
one of these UI failures makes it into the verification process is
overall low. Given the benefits of closing big test gaps, we consider it
worth the risk.

#### Test inclusion

* initially, the regular index files are included
* for this cycle, the feature flag specific UI tests are not included,
this might come as a follow-up