Removes blanket disable of spaces UI features in serverless (#188184)
## Summary

Spaces management and the space selector were previously disabled (not
registered) in serverless, for any and all configurations. This PR will
change the logic for disabling these UI features, only disabling them if
the max spaces count is 1 on serverless. In all other cases (stateful,
and severless with >1 max spaces configured), these features will be
enabled.

This effectively means that at this time, spaces can be fully enabled in
serverless by overriding the setting for max spaces from 1 to anything
>1. E.g. `xpack.spaces.maxSpaces: 1000` (override example
https://github.com/elastic/serverless-gitops/pull/2920).

### Note/Warning
In stateful, the space selector in Kibana's top navigation always
includes a link to manage spaces. In stateful, if a user does not have
privilege to manage spaces, the button will navigate to stack management
by default. In serverless, this behavior is different, and the button
always navigates to the spaces management page, and an error occurs is
the user does not have privilege to manage spaces. This needs to be
addressed before we enable custom spaces in serverless.

Additionally, we will need to determine if we want to feature a
navigation card to the spaces management page in serverless.

### Automated Testing Changes
- `x-pack/plugins/spaces/public/plugin.test.ts`

### Manual Testing

1. Update kibana.dev.yml to include a non-one max spaces setting, e.g. 
```
xpack.spaces.maxSpaces: 1000
```
2. Start ES & Kibana in a local serverless configuration, e.g.
```
yarn es serverless --projectType=es
yarn start --serverless=es
```
3. Log in with the Admin role
4. Confirm the space selector appears in Kibana's top navigation and
behaves as expect
5. Confirm the "Manage space" button navigates to the spaces management
page
6. Repeat test with "Viewer" role
7. Remove the max spaces setting from kibana.dev.yml
8. Confirm the space selector no longer appears in Kibana's top
navigation
9. Start ES & Kibana normally, e.g.
```
yarn es snapshot --license=trial
yarn start
```
10. Confirm that the space selector always appears, regardless of the
max spaces setting