Implements api key creation as admin user in FTR user manager (#188102)
Closes #184948

## Summary

Implements creation of role-scoped API key using admin-scoped
privileges. This ensures that the API keys can be created for any role
scope, and further, that we can use the lowest privileged roles possible
for each API integration test.

## Note to reviewers
Most of the widespread changes here are simply function name changes.

- createApiKeyForRole -> createM2mApiKeyWithRoleScope
- invalidateApiKeyForRole -> invalidateM2mApitKeyWithRoleScope
- getSessionCookieForRole ->
getInteractiveUserSessionCookieWithRoleScope

M2m = Machine to machine. We wanted to be more specific about the
intention - API keys are typically for machine to machine automations
and apps, and cookies are typically for human interactive users.

cc: @azasypkin @dmlemeshko

---------

Co-authored-by: Dzmitry Lemechko <dzmitry.lemechko@elastic.co>