[SecuritySolution][Timeline] Fix issue with creating new timeline (#188538)
## Summary

As shown in https://github.com/elastic/kibana/issues/188500, there is
currently an issue with creating new timelines after visiting the ESQL
tab. It's unclear what has caused this issue but judging by the location
where the error is thrown, this stems from the embedded discover (see
screenshot in issue).

The fix for this issue is to recreate the default discover app state
whenever it is required. After that, no more `read-only` errors are
thrown.

! Please test this fix locally before reviewing !