OTel onboarding: Don't fail on folder existing already (#188657)
The current snippet for Linux and Mac fails if it has been run in the
same working dir already, as the directory the collector is unpacked
into exists already and `mkdir` will fail. This PR adds the `-p` flag to
avoid this.

It also updates the snapshot.

(cherry picked from commit f1eae5c92449332313ceb32189915a3fc5835921)