Rename hook variable to simplify form logic
I had previously used the concept of "some jobs started" because that is
the case that we want to warn users about, and it was conceptually
simpler than "not all jobs started."

However, from both the perspective of the hook and the consumer, it
makes more sense to return an "all jobs started" boolean and use that as
necessary.

This is logically equivalent to the previous commit, just nicer to read.