[Synthetics] allow retest on failure to be turned off (#189013)
## Summary

Resolves https://github.com/elastic/kibana/issues/188804

Allows retest on failure to be configured in the monitor creation UI.

https://www.loom.com/share/2302ccfd150b4668a61966be165478f8

### Testing
1. Create a monitor and toggle retest on failure off.
2. Navigate to edit that monitor. Ensure the setting remains off.

### Release note
Fixes a bug where retest on failure was not able to be turned off when
creating a monitor in the Synthetics app. Retest on failure can now be
turned off.

(cherry picked from commit 14f370b613dbcc5643c5b6e64b7a55938775821a)