fix: [Obs Alerts > Rule detail][SCREEN READER] Multiple single-item lists should be consolidated into one meaningful list: 0001 (#188293)
Closes: https://github.com/elastic/observability-dev/issues/3619

## Description
Observability has a number of description lists that are used to
identify important key-value pairs for users. Visually these look like a
single list of pairs, but they are coded as multiple single item lists.
This is potentially confusing to screen reader users because these data
points are related and should be in a single, consolidated list.

## What was changed?:

1. `RuleDefinition` has been updated to use `EuiDescriptionList`.
2. **The UI was affected.** In the main branch, the `Actions` item is
displayed in a separate column. With `EuiDescriptionList`, it's
displayed with other items.

## Screen: 

<img width="1274" alt="image"
src="https://github.com/user-attachments/assets/4634d44f-2c0b-423f-82a2-5aa04477e142">