feat(investigate): add basic esql widget back (#189092)
Related to https://github.com/elastic/kibana/issues/189091


This PR decouples the NotesWidget and EsqlWidget from the previous
AddWidgetUI, so we can have them both on the page. It's a step forward
the expected design. There is some missing element, like the
visualization selector which I'll add later.
I've also removed the global kuery filter as it is not on the design.

Things not implemented in this PR:
- edit widget
- date range filter (using default 15min from the page)
- lot of cleanup of unused feature

Screenshots |
-- |

![image](https://github.com/user-attachments/assets/791ef79c-1bf8-42b9-87a4-de9fd3f07986)

![image](https://github.com/user-attachments/assets/a326a0ea-a008-4fa7-911f-ea52de318e23)


Testing:
- Run some data forge scripts: `node x-pack/scripts/data_forge.js
--events-per-cycle 50 --lookback now-3d --dataset fake_stack
--install-kibana-assets --kibana-url http://localhost:5601/kibana`
- Go to http://localhost:5601/kibana/app/investigate/new#/
- Create an observation chart based on this esql query: `FROM
kbn-data-forge-fake_stack.admin-console-* | keep @timestamp,
http.response.* | LIMIT 103`

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>
Co-authored-by: Panagiota Mitsopoulou <panagiota.mitsopoulou@elastic.co>