[kbn-test] improve check_ftr_configs script (#189357)
## Summary

Follow-up to #188854

On CI script is taking >1 min, while before it was taking seconds. 
Probably because 80+ files were loaded as potential FTR configs. I
adjusted regular expressions to minimize the amount of files that we
need to load to validate if it is FTR config or not.

In this CI run script took 20s (part of `Quick Checks` group)