[Controls] Fix error on navigation when invalid selections tour step is open (#189449)
## Summary

This PR fixes a bug where, if the "invalid selections" tour step is
active when navigating between dashboards, the dashboard would render
with an `"Unable to load page"` error where the control group should be.
This was because we weren't updating the `invalidSelectionsState` +
`controlWithInvalidSelectionsId` state on navigation.

Note that there is still a race condition happening under the hood
because we are updating the `input` before updating the "invalid
selections" related state - but the `||
!panels[controlWithInvalidSelectionsId]` addition to the control group
component makes it so that this is no longer shown to the user. Since we
will be removing the tour step as part of the control group refactor, it
was not worth digging into this more.

**Before**

https://github.com/user-attachments/assets/d439d52a-1dc8-4a54-b806-85d73fd08099

**After**

https://github.com/user-attachments/assets/541bcc39-41b0-4a8a-ad7b-0de214abe37d

### For maintainers

- [ ] This was checked for breaking API changes and was [labeled
appropriately](https://www.elastic.co/guide/en/kibana/master/contributing.html#kibana-release-notes-process)

(cherry picked from commit b9356a87c207c6295e87f27cdf2417d12c4cfdfd)