[Lens] do not pass incorrect filters to the state (#189292)
## Summary

Before we would pass incorrect filters to the state and save it and then
in the renderer filter whatever we couldn't process.
That unfortunately lead to the bug in another piece of code - when
creating a link to explore underlying data in discover.
This code is aligned with how we do filtering for filtered metric - we
don't pass the filter to the state, unless it is valid.

(cherry picked from commit 2ed5fd9adee317be759a9dec8e326f56b489567b)