[Security Solution][Endpoint] Add authz to file/download apis in support of SentinelOne `processes` response action (#189127)
## Summary

The following changes were done for Response actions:

- The file access APIs (file info + file download) were refactored to
ensure they properly validate the required authz for each type of action
- Now that these APIs are being used for different response actions, we
need to add logic that ensures that a user with access to one response
action (ex. `running-processes` for SentinelOne) is not allowed to
access file information for a different type of action (ex. `get-file`)
- The Response Actions History Log was updated so that the output of a
`processes` response action is now displayed to the user when the
response action row in the table is expanded
- Note that for SentinelOne hosts, the output of the `processes`
response action is a file download - **which will only be visible if
user has authz to Processes operations**
- The height of the expandable row was also increased in order to
provide a larger viewing area for the content that is displayed inside
of it (the action results)