[ResponseOps][alerting] remove debug log causing null dereference on `alert::getUuid()` (#189572)
resolves https://github.com/elastic/kibana/issues/189551

See the reference issue for more details.

We just remove the else clause that is logging the error, as it's both
very unlikely we'd ever see it since it's debug level, but we also don't
think it's needed from the recent work done in this module.

(cherry picked from commit 86bd347e136bfd1463d689999cce628a770bca4c)

# Conflicts:
#	x-pack/plugins/alerting/server/alerts_client/alerts_client.ts