[ResponseOps][TaskManager] `TaskTypeDictionary::get()` no longer `throw`s
resolves https://github.com/elastic/kibana/issues/189560

Removes the `throw` from `TaskTypeDictionary::get()` and have it
return `TaskDefinition | undefined` instead.

Then fix all the places that breaks.  This could probably never happen
when a task is running - the definition would be needed for a bunch
of things prior to this (like polling for the task).  However, it has
occurred during metric generation, and likely more code in the future
will use this, so the new API has less of a "suprise! Error!" aspect
to it, for such an innocent looking method name.

Several places in the existing code did not have to be changed as they
already guards for checking null-ish returns from the method :-)