[Fleet] Retry transient errors in agent status API (#189563)
## Summary

Use `retryTransientEsErrors` when fetching agent status to avoid flaky
ES errors throwing a 500.

I'd love to add tests for this, but it's a bit challenging as these
methods rely heavily on an ES runtime field to populate the status
values. The amount of mocking necessary to add Jest tests for these was
getting a bit ridiculous, so maybe FTR is a better way to go?