chore(NA): skip defend workflows tests failing because of vagrant (#189723)
This is a temporary PR intended to skip the defend workflows FTR suites
that are failing the pipelines due to a vagrant box problem.

It was done in a single PR as it will make it easier to revert after the
problem is fixed.

(cherry picked from commit 3d9d7df6866f9611969e54538ef75267906dabbd)