fix: [ES Serverless home] home page 'Console' button looks like a link (#188518)
Closes: https://github.com/elastic/search-team/issues/7637

## Description

In the 'Select your client' section, 'Console' is a button but it looks
like a link. Screen reader user only hears "Console, button" without
context as to what the user should expect when acting on it. It would be
less confusing if it looked like a button and informed the user that it
will expand the console.

## What was changed?:

1. `TryInConsoleButton` API was slightly updated to support 3 modes of
visualisation: as `link` | `button` | `emptyButton`
2. `ES Serverless home` -> `Console` is now a button

## Screen 

<img width="1219" alt="image"
src="https://github.com/user-attachments/assets/ca1a6be4-2cc5-4a0c-bed9-cf6a1ddeec07">