[Telemetry][Security Solution] Fix flaky tests (#190044)
## Summary

Fixes: https://github.com/elastic/kibana/issues/188234
https://github.com/elastic/kibana/issues/187719 and
https://github.com/elastic/kibana/issues/178918

The flakiness was while calculating the Detection Rules task
invocations. It could have two different RCs: 1) The code didn't retry
in case the task wasn't executed yet, which makes sense in a CI
environment, which is slower than a dev environment; 2) The timestamp to
filter out requests was calculated after the task was triggered, and if
the task is executed fast enough, it could lead to empty responses
because of that.

(cherry picked from commit 358e104ebcc2bfcfeec3457b2e913d071d27745d)