[DOCS] Add more APM UI redirects to the Observability guide (#190061)
## Summary

In https://github.com/elastic/kibana/pull/179981 we moved the APM UI
content to the Observability guide. When [backporting to the `8.14`
branch](https://github.com/elastic/kibana/pull/190009) (which is the
`current` branch at the time I'm writing this) the docs build failed
because there are some links in the APM agent docs that are hard-coded
to go to the `current` version of the docs, and those links were not
properly redirected. When [`8.15` becomes the `current`
branch](https://github.com/elastic/docs/pull/3036), I believe the docs
build will fail unless we implement the AsciiDoc redirects in this PR.

### For maintainers

- [ ] This was checked for breaking API changes and was [labeled
appropriately](https://www.elastic.co/guide/en/kibana/master/contributing.html#kibana-release-notes-process)