[StdPerf] Extending kibana:plugin_render_time with custom metrics (#189115)
## Summary

This PR enable consumers of `PerformanceContextProvider` to track
customMetrics along `onPageReady`.
Tracking customMetrics is important to further analyse and put into
perspective TTFMP. e.g. is not the same trying to load a component that
underneath is using thousand of documents vs one just using a couple of
them.

### Changes
Changes are leveraging existing `PerformanceMetricEvent` schema.

### Ideas
One improvement could be reserving a pair of key/values (aka omitting
them when defining `CustomMetrics` type) for explicitly dedicate them to
timerange values. e.g. the amount of milliseconds between the start and
end date in the timepicker. This could help us to keep the values in the
same property for all observability applications and standardise the
measurement of it.