[Uptime] do not reset pagination when filters have not changed (#189831)
## Summary

Prevents pagination from resetting when Uptime filters are changed.

Previously there was some logic to reset pagination when filters
changed. The intention is to take the user back to page 1 when they
apply a new filter.

However there was not a conditional check against the previous filters
to see if any changes had, in actuality, been applied, and the hook
calling the logic was being called multiple times within the component.

To prevent this unwanted behavior, a conditional was added to check
against the previous state of the filter

### Release note 
Fixes broken pagination in Uptime when a filter is applied. 

### Testing
1. Create 11 monitors, in Synthetics, all with the same tag. The easiest
way to do this is by using `oblt-cli` kibana.yml config, navigating to
Synthetics, creating a monitor with a specific tag, and then copying it
10 times, resulting in 11 monitors.
2. Navigate to Stack Management and toggle uptime on
<img width="1236" alt="Screenshot 2024-08-02 at 4 33 21 PM"
src="https://github.com/user-attachments/assets/aa088bd8-2f73-4349-9b08-3443c0b4f8a4">

3. Refresh the page. Navigate to `app/uptime/settings`. Change the
Uptime index pattern to `synthetics-*`.
4. Navigate to `app/uptime`. In the top filters, select your tag to
apply a filter. There should still be at least 11 monitors to test.
6. Click on the next pagination error. You should be placed on the
second page with only 1 monitor.